backend/app/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
PyJWT==2.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1
qrcode==7.4.2
Pillow==10.0.0 